Fix settings references for Vehicle/Tool/Environment split
- Fixed all Designer.cs files to use correct settings classes
- Fixed type conversions (double to int/decimal) where needed
- Fixed incorrect namespace references (ToolProperties, VehicleProperties)
- Updated CVehicle.cs to use VehicleSettings instead of VehicleProperties
- Updated CAHRS.cs to use Settings.Default for setIMU_isReverseOn
- Updated CTool.cs with cast for minCoverage
- Updated CYouTurn.cs with cast for uTurnSmoothing

All settings now correctly reference:
- Properties.Settings.Default (Environment)
- Properties.VehicleSettings.Default (Vehicle)
- Properties.ToolSettings.Default (Tool)

+++ SourceCode/GPS/Classes/CYouTurn.cs
@@ -99,7 +99,7 @@ namespace AgOpenGPS
 
             uTurnStyle = Properties.Settings.Default.set_uTurnStyle;
 
-            uTurnSmoothing = Properties.Settings.Default.setAS_uTurnSmoothing;
+            uTurnSmoothing = (int)Properties.Settings.Default.setAS_uTurnSmoothing;
         }
 
         //find next not worked lane after the defined lanes to skip
